# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: insights
    description: AI-powered financial insights
  - name: cards
    description: Card onboarding and management
  - name: auth
    description: Authentication and session management
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /insights/health:
    get:
      operationId: insightsHealth
      tags: [insights]
      summary: Insights service health
      description: Returns the AI insights service status and provider info
      responses:
        "200":
          description: Insights service info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InsightsHealthResponse"

  /insights/generate:
    post:
      operationId: generateInsights
      tags: [insights]
      summary: Generate financial insights
      description: Analyzes user financial data and returns AI-powered insights
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InsightsGenerateRequest"
      responses:
        "200":
          description: Generated insights
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InsightsResponse"
        "400":
          description: Invalid request data
        "500":
          description: Insights generation failed

  /cards:
    get:
      operationId: listCards
      tags: [cards]
      summary: List all cards
      responses:
        "200":
          description: Card list
          content:
            application/json:
              schema:
                type: object
                properties:
                  cards:
                    type: array
                    items:
                      $ref: "#/components/schemas/CardRecord"
    post:
      operationId: createCard
      tags: [cards]
      summary: Add a new card
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CardCreateInput"
      responses:
        "201":
          description: Card created
          content:
            application/json:
              schema:
                type: object
                properties:
                  card:
                    $ref: "#/components/schemas/CardRecord"
        "400":
          description: Invalid input

  /cards/{id}:
    get:
      operationId: getCard
      tags: [cards]
      summary: Get a card by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Card details
          content:
            application/json:
              schema:
                type: object
                properties:
                  card:
                    $ref: "#/components/schemas/CardRecord"
        "404":
          description: Card not found
    patch:
      operationId: updateCard
      tags: [cards]
      summary: Update card details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CardUpdateInput"
      responses:
        "200":
          description: Card updated
        "404":
          description: Card not found
    delete:
      operationId: deleteCard
      tags: [cards]
      summary: Remove a card
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Card removed
        "404":
          description: Card not found

  /cards/{id}/freeze:
    post:
      operationId: freezeCard
      tags: [cards]
      summary: Freeze a card
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Card frozen

  /cards/{id}/unfreeze:
    post:
      operationId: unfreezeCard
      tags: [cards]
      summary: Unfreeze a card
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Card unfrozen

  /auth/otp/send:
    post:
      operationId: sendOtp
      tags: [auth]
      summary: Send OTP to phone
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OtpSendRequest"
      responses:
        "200":
          description: OTP sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OtpSendResponse"
        "400":
          description: Invalid phone

  /auth/otp/verify:
    post:
      operationId: verifyOtp
      tags: [auth]
      summary: Verify OTP and authenticate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OtpVerifyRequest"
      responses:
        "200":
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OtpVerifyResponse"
        "401":
          description: Invalid OTP

  /auth/refresh:
    post:
      operationId: refreshToken
      tags: [auth]
      summary: Refresh access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshTokenRequest"
      responses:
        "200":
          description: Token refreshed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RefreshTokenResponse"
        "401":
          description: Invalid refresh token

  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout current session
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LogoutResponse"
        "401":
          description: Unauthorized

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeResponse"
        "401":
          description: Unauthorized

  /auth/devices:
    get:
      operationId: listDevices
      tags: [auth]
      summary: List registered devices
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Device list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeviceListResponse"
    post:
      operationId: registerDevice
      tags: [auth]
      summary: Register a device
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterDeviceRequest"
      responses:
        "201":
          description: Device registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegisteredDevice"
        "400":
          description: Invalid request

  /auth/devices/{id}:
    delete:
      operationId: removeDevice
      tags: [auth]
      summary: Remove a device and revoke its sessions
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Device removed
        "404":
          description: Device not found

  /auth/sessions:
    get:
      operationId: listSessions
      tags: [auth]
      summary: List active sessions
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Session list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionListResponse"

  /auth/sessions/revoke:
    post:
      operationId: revokeSession
      tags: [auth]
      summary: Revoke a session
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RevokeSessionRequest"
      responses:
        "200":
          description: Session revoked
        "400":
          description: Invalid request

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    InsightsHealthResponse:
      type: object
      properties:
        status:
          type: string
        version:
          type: string
        provider:
          type: string
      required:
        - status
        - version
        - provider

    TransactionInput:
      type: object
      properties:
        id:
          type: string
        amount:
          type: number
        type:
          type: string
          enum: [credit, debit]
        category:
          type: string
        description:
          type: string
        date:
          type: string
        status:
          type: string
        merchant:
          type: string
      required:
        - id
        - amount
        - type
        - category
        - description
        - date
        - status
        - merchant

    BudgetInput:
      type: object
      properties:
        id:
          type: string
        category:
          type: string
        limit:
          type: number
        spent:
          type: number
      required:
        - id
        - category
        - limit
        - spent

    ReservedAmountInput:
      type: object
      properties:
        id:
          type: string
        label:
          type: string
        amount:
          type: number
        category:
          type: string
        dueDate:
          type: string
        recurring:
          type: boolean
      required:
        - id
        - label
        - amount
        - category
        - recurring

    InsightsGenerateRequest:
      type: object
      properties:
        transactions:
          type: array
          items:
            $ref: "#/components/schemas/TransactionInput"
        budgets:
          type: array
          items:
            $ref: "#/components/schemas/BudgetInput"
        reservedAmounts:
          type: array
          items:
            $ref: "#/components/schemas/ReservedAmountInput"
        balance:
          type: number
        upiLite:
          type: number
        month:
          type: string
        year:
          type: number
      required:
        - transactions
        - balance

    CategoryBreakdown:
      type: object
      properties:
        category:
          type: string
        total:
          type: number
        count:
          type: integer
        percentage:
          type: number
        trend:
          type: string
          enum: [up, down, stable]
        previousPeriodAmount:
          type: number

    SpendingSummary:
      type: object
      properties:
        totalIncome:
          type: number
        totalSpent:
          type: number
        incomeCount:
          type: integer
        expenseCount:
          type: integer
        averageDailySpend:
          type: number
        largestExpense:
          type: object
          nullable: true
          properties:
            amount:
              type: number
            merchant:
              type: string
            category:
              type: string
        largestIncome:
          type: object
          nullable: true
          properties:
            amount:
              type: number
            merchant:
              type: string
        byCategory:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/CategoryBreakdown"
        topCategory:
          type: string
          nullable: true
        topCategoryAmount:
          type: number

    BudgetRecommendation:
      type: object
      properties:
        category:
          type: string
        limit:
          type: number
        spent:
          type: number
        percentage:
          type: number
        status:
          type: string
          enum: [on_track, warning, exceeded, under_utilized]
        suggestedLimit:
          type: number
        reason:
          type: string

    CashFlowForecast:
      type: object
      properties:
        currentBalance:
          type: number
        projectedBalance:
          type: number
        upcomingExpenses:
          type: array
          items:
            type: object
            properties:
              label:
                type: string
              amount:
                type: number
              dueDate:
                type: string
        daysUntilDepletion:
          type: number
          nullable: true
        surplus:
          type: number
        confidence:
          type: string
          enum: [high, medium, low]

    SavingsOpportunity:
      type: object
      properties:
        category:
          type: string
        currentSpend:
          type: number
        potentialSavings:
          type: number
        suggestion:
          type: string
        impact:
          type: string
          enum: [low, medium, high]

    UnusualTransaction:
      type: object
      properties:
        transactionId:
          type: string
        amount:
          type: number
        merchant:
          type: string
        category:
          type: string
        date:
          type: string
        deviation:
          type: number
        reason:
          type: string
        severity:
          type: string
          enum: [low, medium, high]

    Subscription:
      type: object
      properties:
        name:
          type: string
        amount:
          type: number
        cycle:
          type: string
          enum: [Monthly, Yearly, Weekly, Quarterly]
        category:
          type: string
        merchant:
          type: string
        lastDate:
          type: string
        confidence:
          type: number

    MonthlyTrend:
      type: object
      properties:
        month:
          type: string
        income:
          type: number
        spending:
          type: number
        savings:
          type: number
        topCategory:
          type: string

    InsightRecommendation:
      type: object
      properties:
        icon:
          type: string
        title:
          type: string
        body:
          type: string
        saving:
          type: string
        color:
          type: string
        priority:
          type: string
          enum: [low, medium, high]

    InsightsResponse:
      type: object
      properties:
        healthScore:
          type: number
        healthLabel:
          type: string
        month:
          type: string
        year:
          type: number
        spendingSummary:
          $ref: "#/components/schemas/SpendingSummary"
        budgetRecommendations:
          type: array
          items:
            $ref: "#/components/schemas/BudgetRecommendation"
        savingsOpportunities:
          type: array
          items:
            $ref: "#/components/schemas/SavingsOpportunity"
        unusualTransactions:
          type: array
          items:
            $ref: "#/components/schemas/UnusualTransaction"
        subscriptions:
          type: array
          items:
            $ref: "#/components/schemas/Subscription"
        cashFlowForecast:
          $ref: "#/components/schemas/CashFlowForecast"
        monthlyTrends:
          type: array
          items:
            $ref: "#/components/schemas/MonthlyTrend"
        recommendations:
          type: array
          items:
            $ref: "#/components/schemas/InsightRecommendation"
        generatedAt:
          type: string
        provider:
          type: string
          enum: [local, llm]

    CardRecord:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        cardNetwork:
          type: string
          enum: [visa, mastercard, rupay, amex, discover, unknown]
        issuer:
          type: string
          nullable: true
        lastFour:
          type: string
        expiryMonth:
          type: integer
        expiryYear:
          type: integer
        nickname:
          type: string
        theme:
          type: object
          properties:
            gradientColors:
              type: array
              items:
                type: string
        frozen:
          type: boolean
        balance:
          type: number
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - userId
        - cardNetwork
        - lastFour
        - expiryMonth
        - expiryYear
        - nickname
        - theme
        - frozen
        - balance
        - createdAt
        - updatedAt

    CardCreateInput:
      type: object
      properties:
        cardNetwork:
          type: string
          enum: [visa, mastercard, rupay, amex, discover]
        issuer:
          type: string
          nullable: true
        lastFour:
          type: string
        expiryMonth:
          type: integer
        expiryYear:
          type: integer
        nickname:
          type: string
      required:
        - cardNetwork
        - lastFour
        - expiryMonth
        - expiryYear

    CardUpdateInput:
      type: object
      properties:
        nickname:
          type: string
        frozen:
          type: boolean
        balance:
          type: number

    OtpSendRequest:
      type: object
      required:
        - phone
      properties:
        phone:
          type: string

    OtpSendResponse:
      type: object
      properties:
        message:
          type: string
        expiresIn:
          type: integer

    OtpVerifyRequest:
      type: object
      required:
        - phone
        - otp
      properties:
        phone:
          type: string
        otp:
          type: string
        deviceName:
          type: string
        deviceIdentifier:
          type: string

    AuthUser:
      type: object
      properties:
        id:
          type: string
        phone:
          type: string
        name:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        kycStatus:
          type: string
        twoFactorEnabled:
          type: boolean
        biometricEnabled:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string

    OtpVerifyResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/AuthUser"
        accessToken:
          type: string
        refreshToken:
          type: string

    RefreshTokenRequest:
      type: object
      required:
        - refreshToken
      properties:
        refreshToken:
          type: string

    RefreshTokenResponse:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string

    LogoutResponse:
      type: object
      properties:
        message:
          type: string

    MeResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/AuthUser"

    RegisterDeviceRequest:
      type: object
      required:
        - deviceIdentifier
      properties:
        deviceName:
          type: string
        deviceIdentifier:
          type: string
        pushToken:
          type: string

    RegisteredDevice:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        deviceName:
          type: string
          nullable: true
        deviceIdentifier:
          type: string
        pushToken:
          type: string
          nullable: true
        lastUsedAt:
          type: string
        enrolledAt:
          type: string
        revokedAt:
          type: string
          nullable: true

    DeviceListResponse:
      type: object
      properties:
        devices:
          type: array
          items:
            $ref: "#/components/schemas/RegisteredDevice"

    Session:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        deviceId:
          type: string
          nullable: true
        ipAddress:
          type: string
          nullable: true
        userAgent:
          type: string
          nullable: true
        lastActiveAt:
          type: string
        expiresAt:
          type: string
        revokedAt:
          type: string
          nullable: true
        createdAt:
          type: string

    SessionListResponse:
      type: object
      properties:
        sessions:
          type: array
          items:
            $ref: "#/components/schemas/Session"

    RevokeSessionRequest:
      type: object
      required:
        - sessionId
      properties:
        sessionId:
          type: string